+++ HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ---------------
 
+1.2.2 (2015-04-16)
+++++++++++++++++++
+
+- `@genty_dataprovider` helper methods can now return :class:`GentyArgs`
+  instances to specify args and kwargs.
+
 1.2.1 (2015-04-09)
 ++++++++++++++++++